backend/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "FastAPI backend for the Azure Todo Stack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi~=0.110",
  "uvicorn[standard]~=0.29",
  "sqlmodel~=0.0.16",
  "sqlalchemy~=2.0",
  "psycopg[binary]~=3.1",
  "alembic~=1.13",
  "python-jose[cryptography]~=3.3",
  "python-multipart~=0.0.9",
  "bcrypt~=4.1",
  "structlog~=24.1",
  "pydantic-settings~=2.2",
  "email-validator~=2.1",
  "python-dotenv~=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest~=8.1",
  "pytest-asyncio~=0.23",
  "httpx~=0.27",
  "coverage~=7.4",
  "faker~=24.4"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]
